refactor: 회원탈퇴 로직 수정
