+++ src/app/(auth)/deleteAccount/page.tsx
@@ -7,11 +7,32 @@ import FormTitle from '@/components/common/form/FormTitle';
 import Header from '@/components/common/header/Header';
 import useAuthInput from '@/hooks/useAuthInput';
 import useCheckPassword from '@/queries/auth/useCheckPassword';
+import useModal from '@/hooks/useModal';
+import { useRouter } from 'next/navigation';
+import useGetUser from '@/queries/user/useGetUser';
+import useDeleteAccount from '@/queries/auth/useDeleteAccount';
 
 const DeleteAccountPage = () => {
   const password = useAuthInput({ name: 'password' });
+  const { showModal } = useModal();
+  const router = useRouter();
 
-  const { mutate: checkPassword } = useCheckPassword();
+  const { data: user } = useGetUser();
+  const { mutate: deleteAccount } = useDeleteAccount();
+  const { mutate: checkPassword } = useCheckPassword(() => {
+    showModal(
+      '탈퇴를 진행하시겠습니까?',
+      `계정 탈퇴를 진행하시면,\n${user?.nickname} 님의 여행과 관련된 데이터가\n모두 사라집니다.`,
+      {
+        cancelText: '취소',
+        confirmText: '확인',
+        onCancel: () => {
+          router.push('/mypage');
+        },
+        onConfirm: deleteAccount,
+      },
+    );
+  });
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
