+++ src/queries/auth/useCheckPassword.ts
@@ -3,13 +3,9 @@ import { checkPassword } from '@/api/auth/deleteAccountApi';
 import { QueryError } from '@/@types/query';
 import useModal from '@/hooks/useModal';
 import ModalErrorIcon from '@/assets/modal/modal_error.svg';
-import { useRouter } from 'next/navigation';
-import useDeleteAccount from './useDeleteAccount';
 
-const useCheckPassword = () => {
+const useCheckPassword = (onSuccessCallback: () => void) => {
   const { showModal } = useModal();
-  const router = useRouter();
-  const { mutate: deleteAccount } = useDeleteAccount();
 
   return useMutation({
     mutationFn: checkPassword,
@@ -27,18 +23,7 @@ const useCheckPassword = () => {
       }
     },
     onSuccess: () => {
-      showModal(
-        '탈퇴를 진행하시겠습니까?',
-        `계정 탈퇴를 진행하시면,\n녹차라떼 님의 여행과 관련된 데이터가\n모두 사라집니다.`,
-        {
-          cancelText: '취소',
-          confirmText: '확인',
-          onCancel: () => {
-            router.push('/mypage');
-          },
-          onConfirm: deleteAccount,
-        },
-      );
+      onSuccessCallback();
     },
   });
 };
